Added support for new tsv format and made TsvContent component more dynamic
